Factorisation de journeyController, transfert des méthodes métiers vers la couche service
